fix: stop the annoying ad from showing up every time the nav state change by making the navbar a separate component
